requirements.txt
Django==5.0.10
celery==5.4.0
redis==5.0.1
psycopg2-binary==2.9.9
django-celery-beat==2.6.0
django-celery-results==2.6.0
requests==2.32.3
python-telegram-bot==21.3
djangorestframework==3.15.2
Pillow==10.4.0